pyritship/requirements.txt
# PyRIT Ship Python Dependencies

# Core PyRIT functionality
pyrit

# Web framework (REST API)
flask

# MCP server
mcp

# Environment variable management
python-dotenv

# Optional: Development tools
requests  # Used in request_tester.py